% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{kornbrot_table1}
\alias{kornbrot_table1}
\title{Alertness example data}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{subject}{Subject identifier}
\item{placebo}{The time required to complete a task under the placebo condition}
\item{drug}{The time required to complete a task under the drug condition}
}
}
\source{
\insertRef{kornbrot1990}{rankdifferencetest}
}
\usage{
kornbrot_table1
}
\description{
An example dataset as seen in table 1 from \insertCite{kornbrot1990;textual}{rankdifferencetest}.
The time per problem was recorded for each subject under placebo and drug conditions for the purpose of measuring alertness.
}
\examples{
library(rankdifferencetest)
kornbrot_table1

}
\keyword{datasets}
