% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankdifferencetest-package.r
\docType{package}
\name{rankdifferencetest-package}
\alias{rankdifferencetest}
\alias{rankdifferencetest-package}
\title{rankdifferencetest: Kornbrot's Rank Difference Test}
\description{
Implements Kornbrot's rank difference test as described in <doi:10.1111/j.2044-8317.1990.tb00939.x>. This method is a modified Wilcoxon signed-rank test which produces consistent and meaningful results for ordinal or monotonically-transformed data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://brettklamer.com/work/rankdifferencetest}
}

}
\author{
\strong{Maintainer}: Brett Klamer \email{code@brettklamer.com} (\href{https://orcid.org/0000-0002-8334-4831}{ORCID})

}
\keyword{internal}
