% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srt_pmedian.r
\name{srt_pmedian}
\alias{srt_pmedian}
\title{Calculate the pseudomedian and confidence interval for a signed-rank test.}
\usage{
srt_pmedian(x, ...)
}
\arguments{
\item{x}{(named list)\cr
The object returned by \code{\link[=srt_pvalue]{srt_pvalue()}} in a 'srt pipeline'.}

\item{...}{dots arguments are not used.}
}
\value{
list
}
\description{
Calculates the pseudomedian and confidence interval for a signed-rank test.
}
\references{
\insertRef{exactRankTests}{rankdifferencetest}

\insertRef{R}{rankdifferencetest}
}
\keyword{internal}
