% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srt_pvalue.r
\name{srt_pvalue}
\alias{srt_pvalue}
\title{Calculate the p-value for a signed-rank test.}
\usage{
srt_pvalue(x, ...)
}
\arguments{
\item{x}{(named list)\cr
The object returned by \code{\link[=srt_statistic]{srt_statistic()}} in a 'srt pipeline'.}

\item{...}{dots arguments are not used.}
}
\value{
list
}
\description{
Calculates the p-value for a signed-rank test.
}
\keyword{internal}
