% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srt_result.r
\name{srt_result}
\alias{srt_result}
\title{Prepare the return object for \code{srt()} and \code{rdt()}.}
\usage{
srt_result(x)
}
\arguments{
\item{x}{(named list)\cr
The object returned by \code{\link[=srt_pmedian]{srt_pmedian()}} in an srt pipeline.}
}
\value{
Named \code{list}
}
\description{
Prepares the return object for \code{\link[=srt]{srt()}} and \code{\link[=rdt]{rdt()}}.
}
\keyword{internal}
